requirements.txt
argparse
azure.cosmosCosmosClient
boto3
bs4BeautifulSoup
csv
datetime
dotenvload_dotenv
ioBytesIO
json
os
requests
statistics
sys
time
traceback
xxhash
zipfile